moved barchart files to new folder
